pyproject.toml
[project]
name = "nanopath"
version = "0.1.0"
description = "Lean NanoPath pretraining stack."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "torch==2.8.0+cu129",
  "torchvision==0.23.0+cu129",
  "h5py>=3.16.0",
  "numpy>=2.2.0",
  "pyyaml>=6.0.3",
  "openslide-python>=1.4.0",
  "openslide-bin>=4.0.0",
  "nvidia-ml-py>=12.0.0",
  "wandb>=0.20.0",
]

[tool.uv]
package = false

[tool.uv.sources]
torch = { index = "pytorch-cu129" }
torchvision = { index = "pytorch-cu129" }

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
